feat: coroutine processing job of responsing and loading assets_images (detail_activity)

+++ app/src/main/java/com/codelab/kudongsan/src/main/detail/models/Area.kt
@@ -3,6 +3,6 @@ package com.codelab.kudongsan.src.main.detail.models
 import com.google.gson.annotations.SerializedName
 
 data class Area(
-    @SerializedName("exclusiveArea") val exclusiveArea: Int,
-    @SerializedName("supplyArea") val supplyArea: Int
+    @SerializedName("exclusiveArea") val exclusiveArea: Double,
+    @SerializedName("supplyArea") val supplyArea: Double
 )
