+++ app/src/main/java/com/codelab/kudongsan/src/main/detail/models/Location.kt
@@ -3,6 +3,6 @@ package com.codelab.kudongsan.src.main.detail.models
 import com.google.gson.annotations.SerializedName
 
 data class Location(
-    @SerializedName("latitude") val latitude: Int,
-    @SerializedName("longtitude") val longtitude: Int
+    @SerializedName("latitude") val latitude: Double,
+    @SerializedName("longtitude") val longtitude: Double
 )
